@@ -1,69 +1,41 @@
 {
     "type": "project",
     "license": "proprietary",
-    "minimum-stability": "stable",
-    "prefer-stable": true,
     "require": {
-        "php": ">=8.1",
-        "ext-ctype": "*",
-        "ext-iconv": "*",
+        "php": ">=8.2",
         "ext-pdo": "*",
-        "doctrine/dbal": "^3",
-        "doctrine/doctrine-bundle": "^2.13",
+        "doctrine/dbal": "^3.8",
+        "doctrine/doctrine-bundle": "^2.7",
         "doctrine/doctrine-migrations-bundle": "^3.3",
-        "doctrine/orm": "^3.2",
-        "phpdocumentor/reflection-docblock": "^5.4",
-        "phpstan/phpdoc-parser": "^1.30",
-        "symfony/asset": "6.1.*",
-        "symfony/console": "6.1.*",
-        "symfony/doctrine-messenger": "6.1.*",
-        "symfony/dotenv": "6.1.*",
-        "symfony/expression-language": "6.1.*",
-        "symfony/flex": "^2",
-        "symfony/form": "6.1.*",
-        "symfony/framework-bundle": "6.1.*",
-        "symfony/http-client": "6.1.*",
-        "symfony/intl": "6.1.*",
-        "symfony/mailer": "6.1.*",
-        "symfony/mime": "6.1.*",
-        "symfony/monolog-bundle": "^3.0",
-        "symfony/notifier": "6.1.*",
-        "symfony/process": "6.1.*",
-        "symfony/property-access": "6.1.*",
-        "symfony/property-info": "6.1.*",
-        "symfony/runtime": "6.1.*",
-        "symfony/security-bundle": "6.1.*",
-        "symfony/serializer": "6.1.*",
-        "symfony/string": "6.1.*",
-        "symfony/translation": "6.1.*",
-        "symfony/twig-bundle": "6.1.*",
-        "symfony/validator": "6.1.*",
-        "symfony/web-link": "6.1.*",
-        "symfony/yaml": "6.1.*",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
+        "doctrine/orm": "^3.3",
+        "symfony/asset": "^6.1",
+        "symfony/debug-bundle": "^7.1",
+        "symfony/dotenv": "^7.1",
+        "symfony/framework-bundle": "^6.1",
+        "symfony/mailer": "^7.1",
+        "symfony/messenger": "^7.1",
+        "symfony/monolog-bundle": "^3.10",
+        "symfony/notifier": "^7.1",
+        "symfony/runtime": "^7.1",
+        "symfony/security-bundle": "^7.1",
+        "symfony/translation": "^7.1",
+        "symfony/twig-bundle": "^7.1",
+        "symfony/validator": "^6.1",
+        "symfony/yaml": "^7.1",
+        "twig/extra-bundle": "^3.13",
+        "twig/twig": "^3.1"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
-        "symfony/browser-kit": "6.1.*",
-        "symfony/css-selector": "6.1.*",
-        "symfony/debug-bundle": "6.1.*",
-        "symfony/maker-bundle": "^1.50",
-        "symfony/phpunit-bridge": "^7.1",
-        "symfony/stopwatch": "6.1.*",
-        "symfony/web-profiler-bundle": "6.1.*"
+        "symfony/maker-bundle": "^1.56",
+        "symfony/web-profiler-bundle": "^7.1"
     },
     "config": {
-        "allow-plugins": {
-            "composer/package-versions-deprecated": true,
-            "symfony/flex": true,
-            "symfony/runtime": true
-        },
         "optimize-autoloader": true,
-        "preferred-install": {
-            "*": "dist"
-        },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "symfony/runtime": true
+        }
     },
     "autoload": {
         "psr-4": {
